# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Maicare server API documentation.",
        "title": "Maicare API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/allergy_types": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client_Medical"
                ],
                "summary": "List all allergy types",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_ListAllergyTypesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
        "/clients/{id}/allergies": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client_Medical"
                ],
                "summary": "List all client allergies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-pagination_Response-api_ListClientAllergiesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "client_Medical"
                ],
                "summary": "Create a client allergy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Client allergy data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateClientAllergyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_CreateClientAllergyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CreateClientAllergyRequest": {
            "type": "object",
            "required": [
                "allergy_id",
                "reaction",
                "severity"
            ],
            "properties": {
                "allergy_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "reaction": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                }
            }
        },
        "api.CreateClientAllergyResponse": {
            "type": "object",
            "properties": {
                "allergy_type_id": {
                    "type": "integer"
                },
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "reaction": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                }
            }
        },
...
        "api.ListAllergyTypesResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "api.ListClientAllergiesResponse": {
            "type": "object",
            "properties": {
                "allergy_type": {
                    "type": "string"
                },
                "allergy_type_id": {
                    "type": "integer"
                },
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "reaction": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                }
            }
        },
...
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_CreateClientAllergyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.CreateClientAllergyResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_ListAllergyTypesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.ListAllergyTypesResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-pagination_Response-api_ListClientAllergiesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/pagination.Response-api_ListClientAllergiesResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "pagination.Response-api_ListClientAllergiesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "page_size": {
                    "type": "integer"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ListClientAllergiesResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add 'Bearer ' prefix before your JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}